Reorganize security package
