--- rekolekcje-engine/src/main/java/pl/oaza/warszawa/dor/rekolekcje/api/security/AuthorityName.java
@@ -1,5 +1,0 @@
-package pl.oaza.warszawa.dor.rekolekcje.api.security;
-
-public enum AuthorityName {
-  ROLE_USER, ROLE_ADMIN
-}
