+++ rekolekcje-engine/src/main/java/pl/oaza/warszawa/dor/rekolekcje/api/security/users/AuthorityName.java
@@ -1,0 +1,5 @@
+package pl.oaza.warszawa.dor.rekolekcje.api.security.users;
+
+public enum AuthorityName {
+  ROLE_USER, ROLE_ADMIN
+}
