+++ rekolekcje-engine/src/main/java/pl/oaza/warszawa/dor/rekolekcje/api/security/users/UserRepository.java
@@ -1,8 +1,8 @@
-package pl.oaza.warszawa.dor.rekolekcje.api.security;
+package pl.oaza.warszawa.dor.rekolekcje.api.security.users;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import pl.oaza.warszawa.dor.rekolekcje.api.security.User;
+import pl.oaza.warszawa.dor.rekolekcje.api.security.users.User;
 
 @RepositoryRestResource(exported = false)
 public interface UserRepository extends JpaRepository<User, Long> {
